graph navigation will now use config to avoid stairs as necessary
